renames longnet file; longnet example in readme works now
