+++ examples/longvit/longvit.py
@@ -23,7 +23,7 @@ import torch
 import torch.nn as nn
 
 from torchscale.architecture.encoder import Encoder
-from torchscale.model.LongNet import LongNetEncoder
+from torchscale.model.longnet import LongNetEncoder
 from torchscale.architecture.config import EncoderConfig
 from timm.models.layers import trunc_normal_ as __call_trunc_normal_
